feat: Enhance performance dashboard error reporting by parsing detailed backend messages, improving UI display, and ensuring correct earnings type in the database.

+++ supabase/migrations/0006_performance_tiers.sql
@@ -85,7 +85,7 @@ BEGIN
     WITH earnings AS (
         SELECT 
             c.talent_id,
-            SUM(c.talent_earnings_cents) as total_earnings
+            SUM(c.talent_earnings_cents)::BIGINT as total_earnings
         FROM public.campaigns c
         WHERE c.agency_id = p_agency_id
           AND c.status = 'Completed'
